requirements_dev.txt
torch
matplotlib
torchvision
pandas
Pillow
requests
kaggle